preliminary fastapi converting
