+++ src/api/requirements.txt
@@ -7,3 +7,5 @@ flask_caching==1.9.0
 alembic==1.4.3
 pytest==6.2.5
 requests==2.26.0
+uvicorn==0.14.0
+pydantic==
